+++ src/Core/WorkSpace/src/Components/DataViewerHeader/Header.tsx
@@ -15,6 +15,7 @@ import { useViewerContext } from '../../Context/ViewProvider';
 import { useIntervalTimestamp } from '../../Hooks/useIntervalTimestamp';
 import { TabsConfigItem } from '../../Util/tabsConfig';
 import { Presets } from '../Presets/Presets';
+import { SearchButton } from '../Search/Search';
 import { TabButton } from '../ToggleButton';
 import {
     ActionBar,
@@ -84,6 +85,7 @@ export const CompletionViewHeader = ({
                             {pages.map((page) => {
                                 return (
                                     <TabButton
+                                        width={`${page.pageTitle.length * 10}px`}
                                         aria-selected={
                                             (activePage?.pageId &&
                                                 page.pageId === activePage.pageId &&
@@ -106,12 +108,12 @@ export const CompletionViewHeader = ({
                     {factory && (
                         <>
                             <TabButton
+                                width={'48px'}
                                 onClick={factory.onClick}
                                 aria-selected={false}
                                 title={factory.title}
                             >
                                 <Icon name={'add'} />
-                                {factory.title}
                             </TabButton>
                             <Divider />
                         </>