+++ src/apps/DisciplineReleaseControl/Components/Form/Inputs/index.ts
@@ -1,4 +1,0 @@
-export * from './Description/Description';
-export * from './Title/Title';
-export * from './PlannedStartDate/PlannedStartDate';
-export * from './References/References';