+++ src/apps/DisciplineReleaseControl/WorkspaceConfig/index.ts
@@ -1,4 +1,3 @@
-export * from './DataCreator/dataCreatorConfig';
 export * from './Filter/filterConfig';
 export * from './Garden/gardenConfig';
 export * from './Presets/presetConfig';