+++ src/apps/ReleaseControl/api/releaseControl/Request/index.ts
@@ -3,4 +3,3 @@ export * from './getReleaseControlById';
 export * from './postReleaseControl';
 export * from './voidRequest';
 export * from './attachment';
-export * from './initiateReleaseControl';