+++ src/apps/ReleaseControl/workspaceConfig/garden/gardenConfig.ts
@@ -4,8 +4,9 @@ import { ReleaseControl } from '../../types/releaseControl';
 export const gardenOptions: GardenOptions<ReleaseControl> = {
     gardenKey: 'Status' as keyof ReleaseControl,
     itemKey: 'sequenceNumber',
-    type: 'normal',
+    objectIdentifier: 'id',
     collapseSubGroupsByDefault: true,
+    customDescription: ({ title }) => title,
     fieldSettings: {
         CurrentStep: {
             getKey: ({ currentWorkflowStep }) => currentWorkflowStep?.name ?? '(Blank)',