+++ src/apps/ScopeChangeRequest/workspaceConfig/sTable/tableConfig.tsx
@@ -10,9 +10,12 @@ import { Atom, deref, swap } from '@dbeining/react-atom';
 import { EstimateBar } from '../../Components/WoProgressBars/EstimateBar';
 import { ExpendedProgressBar } from '../../Components/WoProgressBars/ExpendedProgressBar';
 import styled from 'styled-components';
+import { TableData, CellRenderProps, CellProps } from '@equinor/Table';
 
 const customCellView = (render: (req: ScopeChangeRequest) => JSX.Element | null) => ({
-    Cell: ({ cell }: any) => <>{render(cell.value.content)}</>,
+    Cell: (
+        cell: React.PropsWithChildren<CellProps<TableData, CellRenderProps<ScopeChangeRequest>>>
+    ) => <>{render(cell.value.content)}</>,
 });
 
 export const tableConfig: TableOptions<ScopeChangeRequest> = {
@@ -47,6 +50,25 @@ export const tableConfig: TableOptions<ScopeChangeRequest> = {
                 return <div>{lastSigned.toRelative({ locale: 'en-GB' })}</div>;
             },
             id: 'LastSigned',
+            width: 180,
+            Aggregated: () => null,
+            aggregate: 'count',
+        },
+        {
+            Header: 'Disciplines',
+            accessor: 'disciplineGuesstimates',
+            Cell: ({ cell }: any) => {
+                const request = cell.row.original as ScopeChangeRequest;
+
+                return (
+                    <div>
+                        {request.disciplineGuesstimates
+                            .map(({ discipline: { procosysCode } }) => procosysCode)
+                            .toString()}
+                    </div>
+                );
+            },
+            id: 'Disciplines',
             width: 180,
             Aggregated: () => null,
             aggregate: 'count',
@@ -88,6 +110,7 @@ export const tableConfig: TableOptions<ScopeChangeRequest> = {
         'actualChangeHours',
         'changeCategory',
         'originSource',
+        'scope',
         'modifiedAtUtc',
         'systems',
         'areas',
@@ -101,6 +124,7 @@ export const tableConfig: TableOptions<ScopeChangeRequest> = {
         { key: 'sequenceNumber', title: 'Id', width: 60 },
         { key: 'title', title: 'Title', width: 250 },
         { key: 'phase', title: 'Phase', width: 60 },
+        { key: 'id', title: 'GUID' },
         { key: 'workflowSteps', title: 'Workflow', width: 110 },
         { key: 'disciplineGuesstimates', title: 'Guess mhrs', width: 120 },
         { key: 'estimatedChangeHours', title: 'Est mhrs', width: 120 },
@@ -133,6 +157,15 @@ export const tableConfig: TableOptions<ScopeChangeRequest> = {
             key: 'scope',
             type: customCellView((req) => <>{req?.scope?.name}</>),
         },
+        {
+            key: 'createdBy',
+            type: customCellView(
+                (req) =>
+                    req.createdBy && (
+                        <>{`${req?.createdBy?.firstName} ${req?.createdBy?.lastName}`}</>
+                    )
+            ),
+        },
 
         {
             key: 'createdAtUtc',
@@ -160,7 +193,7 @@ export const tableConfig: TableOptions<ScopeChangeRequest> = {
         {
             key: 'disciplineGuesstimates',
             type: {
-                Cell: ({ cell }: any) => {
+                Cell: ({ cell }: any): JSX.Element => {
                     const request: ScopeChangeRequest = cell.value.content;
 
                     if (deref(guesstimateHoursMaxAtom) === -1) {
@@ -263,6 +296,24 @@ export const tableConfig: TableOptions<ScopeChangeRequest> = {
                     )}
                 </>
             )),
+        },
+        {
+            key: 'materialsToBeBoughtByContractor',
+            type: customCellView((req) => (
+                <>{req.materialsToBeBoughtByContractor ? 'Yes' : 'No'}</>
+            )),
+        },
+        {
+            key: 'materialsIdentifiedInStorage',
+            type: customCellView((req) => <>{req.materialsIdentifiedInStorage ? 'Yes' : 'No'}</>),
+        },
+        {
+            key: 'potentialWarrantyCase',
+            type: customCellView((req) => <>{req.potentialWarrantyCase ? 'Yes' : 'No'}</>),
+        },
+        {
+            key: 'isVoided',
+            type: customCellView((req) => <>{req.isVoided ? 'Yes' : 'No'}</>),
         },
         {
             key: 'workflowSteps',