+++ src/apps/WorkOrder/Garden/components/WorkOrderItem/WorkOrderItem.tsx
@@ -1,4 +1,4 @@
-import { FlagIcon, ProcosysStatuses, FollowUpStatuses } from '@equinor/GardenUtils';
+import { FlagIcon, ProcosysStatuses, FollowUpStatuses, PopoverWrapper } from '@equinor/GardenUtils';
 import { memo, useMemo, useRef, useState } from 'react';
 import { useParkViewContext } from '../../../../../components/ParkView/Context/ParkViewProvider';
 import { CustomItemView } from '../../../../../components/ParkView/Models/gardenOptions';
@@ -59,10 +59,14 @@ const WorkOrderItem = ({
     onClick,
     columnExpanded,
     depth,
-    selectedItem,
+    isSelected,
+    columnStart,
+    parentRef,
+    rowStart,
     width: itemWidth = 300,
 }: CustomItemView<WorkOrder>) => {
     const [isOpen, setIsOpen] = useState<boolean>(false);
+    const [hoverTimeout, setHoverTimeout] = useState<ReturnType<typeof setTimeout> | null>(null);
     const anchorRef = useRef<HTMLDivElement>(null);
     const { groupByKeys, gardenKey } = useParkViewContext<WorkOrder>();
 
@@ -82,36 +86,27 @@ const WorkOrderItem = ({
 
     const width = useMemo(() => (depth ? 100 - depth * 3 : 100), [depth]);
     const maxWidth = useMemo(() => itemWidth * 0.98, [itemWidth]);
+
     return (
         <>
-            {/* {isOpen && (
-                <WorkOrderPopover
-                    data={data}
-                    anchorRef={anchorRef}
-                    isOpen={isOpen}
-                    itemOptions={{
-                        barColor: backgroundColor,
-                        textColor: textColor,
-                        milestone: status,
-                        size,
-                        matStatus,
-                        matColor,
-                        mccrColor,
-                    }}
-                />
-            )} */}
             <Root>
                 <WorkOrderWrapper
                     backgroundColor={backgroundColor}
                     textColor={textColor}
                     background={progressBar}
                     ref={anchorRef}
-                    onMouseOver={() => setIsOpen(true)}
-                    onMouseLeave={() => setIsOpen(false)}
+                    onMouseEnter={() => {
+                        hoverTimeout && !isOpen && clearTimeout(hoverTimeout);
+                        setHoverTimeout(setTimeout(() => setIsOpen(true), 1000));
+                    }}
+                    onMouseLeave={() => {
+                        hoverTimeout && clearTimeout(hoverTimeout);
+                        setIsOpen(false);
+                    }}
                     onClick={onClick}
                     style={{ width: `${columnExpanded ? 100 : width}%`, maxWidth }}
                     progressBackground={progressBar}
-                    isSelected={selectedItem?.workOrderNumber === data.workOrderNumber}
+                    isSelected={isSelected}
                 >
                     <Sizes size={size} color={textColor} />
                     {data.holdBy && <FlagIcon color={textColor} />}
@@ -120,6 +115,29 @@ const WorkOrderItem = ({
                 </WorkOrderWrapper>
                 {columnExpanded && data.description}
             </Root>
+            {isOpen && (
+                <PopoverWrapper
+                    popoverTitle={`Wo.Number: ${data.workOrderNumber}`}
+                    width={itemWidth}
+                    columnStart={columnStart}
+                    parentRef={parentRef}
+                    rowStart={rowStart}
+                    isOpen={isOpen}
+                >
+                    <WorkOrderPopover
+                        data={data}
+                        itemOptions={{
+                            barColor: backgroundColor,
+                            textColor: textColor,
+                            milestone: status,
+                            size,
+                            matStatus,
+                            matColor,
+                            mccrColor,
+                        }}
+                    />
+                </PopoverWrapper>
+            )}
         </>
     );
 };