+++ src/apps/WorkOrder/index.tsx
@@ -33,13 +33,15 @@ export function setup(appApi: ClientApi): void {
             responseParser: responseParser,
         })
         .registerFilterOptions(filterConfig)
+        .registerSearchOptions([
+            { name: 'Id', valueFormatter: ({ workOrderNumber }) => workOrderNumber },
+        ])
         .registerTableOptions(tableConfig)
         .registerGardenOptions({
             gardenKey: 'fwp' as keyof WorkOrder,
             itemKey: 'workOrderNumber',
             fieldSettings: fieldSettings,
-
-            type: 'virtual',
+            objectIdentifier: 'workOrderId',
             customViews: {
                 customItemView: WorkOrderItem,
                 customHeaderView: WorkOrderHeader,