+++ src/packages/Table/index.ts
@@ -1,5 +1,5 @@
 export * from './Components/Table';
 export * from './Components/Cells';
 export * from './Hooks';
-export * from './types';
+export * from './Types';
 export { defaultGroupByFn } from './Utils/utils';